+++ frontend/package.json
@@ -20,6 +20,7 @@
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
+    "react-images-uploading": "^3.1.7",
     "react-modal": "^3.16.1",
     "react-number-format": "^5.4.0",
     "react-redux": "^9.1.0",
